% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr6-package.R
\docType{package}
\name{minidplyr6-package}
\alias{minidplyr6}
\alias{minidplyr6-package}
\title{minidplyr6: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
