% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{select2}
\alias{select2}
\title{Select some column}
\usage{
select2(df, names_or_ind)
}
\arguments{
\item{df}{a data frame.}

\item{names_or_ind}{vector of column name or vector of column indicies.}
}
\value{
a data frame with the subsetted columns
}
\description{
Select some column
}
